Refactors the integration test scripts
* Creates a new script to start a development server, previously
  generating test tilesets.

* Refactors the integration test infrastructure to extract the tileset
  generation and server startup out of the integration test themselves.
  This allows us to start a server and generate the tilesets for
  development purposes.
