+++ Vagrantfile
@@ -4,10 +4,10 @@ Vagrant.configure("2") do |config|
   config.vm.box = "ubuntu/precise64"
 
   # Forward server port
-  config.vm.network :forwarded_port, host: 8080, guest: 8080, auto_correct: true
+  config.vm.network :forwarded_port, host: 8000, guest: 8000, auto_correct: true
 
   # Provision the development environment
-  config.vm.provision :shell, path: "Vagrantprov", privileged: false
+  config.vm.provision :shell, path: "Vagrantprov", privileged: true
 
   # Add more virtual memory
   config.vm.provider "virtualbox" do |v|
